df_mlfit1 works
The c sources have been fixed in various places and more unit tests have
been placed in df_mlfit1_ut. This fitting routine pretty much the same
as what is found in dotFitting.m with the major difference is that it is
written in c using gsl. Things to fix before swithing to this
implementation:

 * Allow specification of sigma per dimension.
 * Use sub pixel precision initial positions together with comw for
faster convergence.
 * Integrate a filter to throw away fittings that drifted away, i.e.,
dot that moved more than 1 unit in any dimensions.