Crash fix and improvements
Fixed a crash that occured when the new dot detection pipeline was used
together with FWHM calculations.

Fixed a crash that occured when there were no nuclei at all.

Switched to the new dot detection in the CC correction pipeline.

Added toggle switches for individual channels when viewing the picked
dots in the CC generation.