Improved stability of cc-correction
The algorithm for cc-correction did fail in certain cases resulting in
NaNs. This problem was easy to spot and was of type fail or work. No
actions required for old data. Hopefully it will work with more data
sets now. Thanks to SW for sensing that something was wrong.

Also fixed the issue where the same field of view would be listed
multiple times.