Fixes involuntarily rounding
Seems like writetable has a whimsical behaviour and sometimes write a
complete column as integers just because a few of the values were
integers. The solution, i.e., df_writetable is not as flexible as
writetable and much slower. But the behaviour is more like it should be.
Using three decimals as default.