Save saving
df_nm_save should be more safe to write than saving directly with 'save'
it first writes to a temprary file and then moves the temporary file to
the target file name. By doing that it should be virtually impossible to
write corrupt files (i.e., should not matter if MATLAB crashes while
writing because then the original file will not be changed).